requirements.txt
Flask
docx2txt
python-docx
Flask-Cors
